created app for deployment

+++ app/controllers/events_controller.rb
@@ -1,33 +1,59 @@
 class EventsController < ApplicationController
+    before_action :require_login, only: [:edit, :update, :destroy]
+
     def index
-        @in_state_events = Event.where(state: current_user.state)
-        @other_events = Event.where.not(state: current_user.state)
-        render "/events/index.html.erb"
+        user = User.find(current_user.id)
+        @local_events = Event.where(state: user.state)
+        @other_events = Event.where.not(state: user.state)
+    end
+
+    def show
+        event = Event.find(params[:id])
+        user = event.users
+        comments = event.comments
     end
 
     def create
-        event = Event.create(user_params)
+        event = Event.create(event_params)
 
-        if event.valid?
-            flash[:errors] = ["You have successfully created an event!"]
-            Participant.create(user: current_user, event: event)
+        if event.save
             redirect_to "/events"
-
         else
            flash[:errors] = event.errors.full_messages
-           redirect_to root_path
+           redirect_to :back
        end
-    end
-
-    def view
-        @events = Event.find(params[:id])
-        @comments = Comment.where(event: @event)
-        render "/events/view.html.erb"
     end
 
     def destroy
         event = Event.find(params[:id])
-        event.destroy
+        event.destroy if event.user == current_user
         redirect_to "/events"
     end
+
+    def edit
+        event = Event.find(params[:id])
+        event.destroy if event.user == current_user
+        redirect_to "/events"
+    end
+
+    def update
+        event = Event.find(params[:id])
+        if event.update(event_params)
+            redirect_to "/events/#{@event.id}"
+        else
+            flash[:errors] = event.errors.full_messages
+            redirect_to :back
+    end
+
+    private
+        def event_params
+            params.require(:event).permit(:name, :date, :location, :state).merge(user: current_user)
+        end
+
+        def require_login
+            user_id = params[:id].to_i
+
+            if user_id != session[:user_id]
+                redirect_to "/events"
+        end
 end
